init: start react-game task
feat: make basic app structure
feat: add button control for char
feat: jumping for char
feat: add obstacles generation and motions
feat: add navigation bar (non-func)
feat: add characters.js - for safe characters and obstacles extension
