+++ src/assets/chars/soldier/sprite.png
